fix paging source nya

+++ core/src/main/java/com/example/mymovie/core/data/source/remote/response/MovieResponse.kt
@@ -35,7 +35,7 @@ data class MovieResponse(
 	val popularity: Double? = null,
 
 	@field:SerializedName("vote_average")
-	val voteAverage: Double? = null,
+	val voteAverage: Float? = null,
 
 	@field:SerializedName("id")
 	val id: Int? = null,
